chaintsn: started refactoring for epoch-level bookkeeping
