+++ crates/chaintsn/src/lib.rs
@@ -1,5 +1,6 @@
 //! Rollup chain state transition.
 
+pub mod epoch;
 pub mod errors;
 pub mod slot_rng;
 pub mod transition;
